refactor: move argument handling and configuration to its own module
